refactor: extract terminal / ui_state / github modules out of main.rs
main.rs にべた書きだった PTY 起動・Slint モデル詰め替え・EventProxy
定義等を以下のモジュールに分割し、main は composition root のみに
絞る。既存の Terminal ペイン挙動は変えない。

- src/terminal/ : PTY + alacritty_terminal + 子プロセス管理 + key translate
- src/ui_state/ : Slint の TerminalRow / TerminalCell への詰め替え
- src/github/   : 将来の octocrab 連携のための空モジュール
- src/main.rs   : argv パース → terminal::launch(&ui, cmd) → ui.run() のみ

translate_key / EventProxy / TermSize / empty_row / build_row / make_cell
は terminal または ui_state に移動。挙動は完全に同じ。

Refs: #200
