+++ src/github/mod.rs
@@ -1,0 +1,8 @@
+#![allow(dead_code)]
+
+//! GitHub 連携のための境界モジュール。
+//!
+//! v0.1 (#200) の段階では空。次 Issue (#201) で octocrab ラッパと
+//! `IssueContextProvider` trait を追加する。意図的に先にモジュールだけ
+//! 切っておき、main.rs が依存関係の変更なしに次の変更を受け入れられる
+//! ようにしておく。
